Independs methods which related to "digest"
